refactor: extract settings filter bar
